eduAgent/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
email-validator
pymongo
python-dotenv
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
qdrant-client